remove(ng): merge sessions_list_ctrl into sessions_map_ctrl
