--- app/javascript/angular/code/controllers/sessions_list_ctrl.js
@@ -1,12 +1,0 @@
-import { SessionsListCtrl } from "./_sessions_list_ctrl";
-
-angular
-  .module("aircasting")
-  .controller("SessionsListCtrl", [
-    "$scope",
-    "params",
-    "$window",
-    "sessionsUtils",
-    "map",
-    SessionsListCtrl
-  ]);
