+++ app/javascript/angular/tests/_sessions_list_ctrl.test.js
@@ -1,6 +1,6 @@
 import test from "blue-tape";
 import { mock } from "./helpers";
-import { SessionsListCtrl } from "../code/controllers/_sessions_list_ctrl";
+import { SessionsMapCtrl } from "../code/controllers/_sessions_map_ctrl";
 
 test("with no sessions selected and isSearchAsIMoveOn true when params.map changes it calls sessions.fetch", t => {
   const callbacks = [];
@@ -14,7 +14,7 @@ test("with no sessions selected and isSearchAsIMoveOn true when params.map chang
   const params = {
     get: () => ({ isSearchAsIMoveOn: true })
   };
-  _SessionsListCtrl({ $scope, sessions, params });
+  _SessionsMapCtrl({ $scope, sessions, params });
 
   callbacks.forEach(callback =>
     callback({ hasChangedProgrammatically: false })
@@ -34,7 +34,7 @@ test("with session selected when params.map changes it does not call sessions.fe
   const sessions = {
     ...mock("fetch")
   };
-  _SessionsListCtrl({ $scope, sessions });
+  _SessionsMapCtrl({ $scope, sessions });
 
   callbacks.forEach(callback => callback({}));
 
@@ -43,16 +43,16 @@ test("with session selected when params.map changes it does not call sessions.fe
   t.end();
 });
 
-const _SessionsListCtrl = ({
-  map,
-  $scope,
-  updateCrowdMapLayer,
-  sessions,
-  params
-}) => {
-  const _sessions = { reSelectAllSessions: () => {}, ...sessions };
+const _SessionsMapCtrl = ({ $scope, sessions, params }) => {
+  const _sensors = {
+    setSensors: () => {}
+  };
+  const $window = {};
+  const _sessions = {
+    isMobile: () => false,
+    ...sessions
+  };
   const _$scope = {
-    sessions: _sessions,
     setDefaults: () => {},
     $watch: () => {},
     $on: () => {},
@@ -60,19 +60,22 @@ const _SessionsListCtrl = ({
   };
   const _params = {
     get: () => ({}),
-    update: () => {},
-    paramsData: {},
+    updateFromDefaults: () => {},
     ...params
   };
   const _map = {
-    onPanOrZoom: () => {},
-    ...map
-  };
-  const _updateCrowdMapLayer = updateCrowdMapLayer || {};
-  const $location = {
-    path: () => "/map_sessions"
+    unregisterAll: () => {}
   };
   const _sessionsUtils = { isSessionSelected: () => false };
 
-  return SessionsListCtrl(_$scope, _params, {}, _sessionsUtils, _map);
+  return SessionsMapCtrl(
+    _$scope,
+    _params,
+    _map,
+    _sensors,
+    _sessions,
+    null,
+    $window,
+    _sessionsUtils
+  );
 };
